Movement/Jumping II
Two modes of physics movement are implemented and functioning for walking, running, jumping, and double-jumping. "Direct velocity" and "force."

Animations are implemented for idle, walking, running, and jumping, and are velocity-based.

The number of extra jumps can be specified as well as two separate movement profiles - one for each type of movement mode.

The movement modes are switchable with a boolean, and can be combined as the level designer sees fit, with events that toggle the flags.

Next up are parameters to control the falling trajectory and crouching.