Movement/Jumping IV, Jetpack II, Ranged Weapons I
Added gravity and descent controls to force based jumping and changed some calculations to allow restricting the max jump velocity.

Made the jetpack thrust visible regardless of parent object sprite normals.

Added animations for aiming the gun up and down while running, and set up the logic to fire the gun correctly.

Created a projectile prefab and configured it for the player.

Added all the art assets for the environments and actors that I have received so far.