Movement/Jumping V, Camera II, Debris I
Added the ability to toggle off walking, forcing running at all times. Also, added the stationary animation for shooting up, down, and to the side, added the logic, and made sure that they work correctly.

Activated sprites for the jetpack exhaust.

Added elasticity and limits to camera.

Created a debris marker that drops particles and will instantiate debris (later). Hooked it into an 'Intensity' mechanic that is governed by the GameManager, based on the player's rate of ascent (and other triggers), so that it will increase logarithmically based on intensity.