Camera III, Swarmling II, Puzzle Elements II, Editor Aid I
Implemented camera shaking based off of the intensity variable in the GameManager, and fully configurable in the editor.

Set up the new Swarmling animations. Turned Swarmling NavPoints into spawn points and added a system to the ActorManager that increases or decreases the number of Swarmlings based on intensity in the GameManager. It will only spawn them at Swarmling NavPoints in the goldilocks zone beneath the player and will remove them once they are past the minimum distance of the goldilocks zone.

Added the Door animations and tested them.

Set up Gizmos for the NavPoints to show where they are and what the NavPoint path is for Swarmlings during  the level design phase.

Added all of the art Brosnan provided.