Refactoring I, Swarmling IV, BugFix I
Split a Damager class out of the ActorController class that handles one melee attack and/or one ranged attack.

Added a rebounding behavior to melee attacks that allows them to collide again after a moment.

Adjusted the doors so that they have proper collision with projectiles, and adjusted the ActorController so that it doesn't allow shooting when dead.