Puzzle Elements III, Cyborg I
Added the two new door variants.

Added the Cyborg and set up his character controller and AI, and modified the NavPoints so that a non-Swarmling AI Actor can follow any set of NavPoints. With a little modification, this setup can be used as floating platforms.

Broke the Swarmling's targeting behavior.